requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
ollama>=0.1.0
ipython>=8.0.0